A few changes :^)
One of the changes is a fix for a problem with the post-processing
done on the downloaded file file. The problem was that `yt-dlp` was not
telling us where the file would ctually be saved, e.g. when downloading
an audio file with the parameter `-x`, the filename reported by `yt-dlp`
would still end with `webm` (`opus` expected). Therefore, we now remove
the video track ourselves if the user requests audio only.

Another notable change is controlling where `yt-dlp` downloads the
file. Before it would download the unprocessed file into the current
directory, and we wouldn't remove it. Therefore, we now use the
`directories` crate to get the users `Downloads` directory, set `yt-
dlp`'s `cwd` to `/tmp` (or the equivalent on a non-unix platform), and
then process the file and output it into `Downloads`.

Details:
- Added the `directories` crates as a dependency to the project
- Changing the `filepath` member of the struct `RequestedDownload` to `filename` since we now control the
download dir
- Add the `download_dir` member to the `Config` struct so that we can download the file to the right place
- Added the `make_base_command()` function to create the base command
for `yt-dlp` since we now have a seperate function to download the file
and get the information from `yt-dlp` (should I have mentioned that
before?)
- Other small details which are related to previous changes, if anyone is actually interested they can probably find them themselves :)

Signed-off-by: Erick Howard <needlesslygrim@proton.me>