Refactor: Move the `resolution` field of `Config` into `MediaType::Video`
This means that we don't need to have an `Option<Resolution>`, and we
can just extract the resolution easily when matching `media_type`. It
also makes the most sense from a logical point of view. The
implementation of `TryFrom<usize>` has been removed for now since
it doesn't make sense with that behaviour now.

Signed-off-by: Erick Howard <needlesslygrim@proton.me>