Very basic initial version of the GUI
Written in `egui` with `eframe`. Uses `rfd` for the file picker. Doesn't
download anything but that should be possibly maybe not definitely easy
to implement.

Conditional compilation with a Cargo feature is utilised to avoid
pulling in large dependencies for the CLI version, `yabw::run()` will
either have the previous definition, or the new definition which calls
`yabw::App::build()` to build the UI.

Other details:
- Cleaned up `main.rs`
- Made some structs and enums public, and some

Signed-off-by: Erick Howard <needlesslygrim@proton.me>